final-pjt-back/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
dj-rest-auth==5.0.2
Django==4.2.5
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.6
